# display-video/openapi.yaml
openapi: 3.0.3
info:
  title: Display & Video Campaign Manager API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /v1/ads:
    get:
      summary: List ads
      parameters:
        - name: assetGroupId
          in: query
          schema: { type: string }
    post:
      summary: Create ad (file-based, multipart)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [assetGroupId, name, adType, inputType, file]
              properties:
                assetGroupId: { type: string }
                name: { type: string }
                adType: { type: string, enum: [DISPLAY, VIDEO] }
                inputType: { type: string, enum: [DISPLAY_IMAGE, DISPLAY_HTML5_ZIP, VIDEO_FILE] }
                landingUrl: { type: string }
                brandUrl: { type: string }
                sponsoredBy: { type: string }
                ctaText: { type: string }
                file: { type: string, format: binary }
...